pyproject.toml
[build-system]
requires = ["uv_build>=0.11.2,<0.12"]
build-backend = "uv_build"

[project]
name = "fin-ai-stats-extract"
version = "0.1.2"
description = "Extract structured AI investment data from earnings call transcripts using OpenAI"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=1.82.0",
    "pydantic>=2.11.0",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.7.0",
    "tqdm>=4.67.0",
]

[project.scripts]
fin-ai-stats-extract = "fin_ai_stats_extract.cli:main"
